It6 frequent_destination_name / date_with_most_trips

+++ db/migrate/20171004220124_create_trips.rb
@@ -8,7 +8,7 @@ class CreateTrips < ActiveRecord::Migration[5.1]
       trip.integer   :end_station_id
       trip.integer   :bike_id
       trip.text      :subscription_type
-      trip.integer   :zip_code
+      trip.bigint  :zip_code
     end
   end
 end
